refactor: 전체선택 로직, 선택된 장바구니 로직 분리
