+++ src/hooks/useSelectedItems.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useRecoilState } from 'recoil';
 import { SelectedCartItem, selectedCartItems } from '../recoil/atoms';
 import intersectionByProperty from '../utils/getArrayIntersection';
@@ -9,7 +9,6 @@ const useSelectedItems = (
   getOneItemQuantity: (id: number) => number | undefined,
 ) => {
   const [selectedItems, setSelectedItems] = useRecoilState(selectedCartItems);
-  const [all, setAll] = useState<boolean>(data.length === selectedItems.length);
 
   useEffect(() => {
     if (data.length !== 0) {
@@ -28,10 +27,6 @@ const useSelectedItems = (
       });
     }
   }, [data, setSelectedItems]);
-
-  useEffect(() => {
-    setAll(data.length === selectedItems.length);
-  }, [data, selectedItems]);
 
   // 현재 selectedItems에서 id가 있는지에 대한 함수
   // id를 받아서 id가 있는지를 boolean
@@ -62,24 +57,6 @@ const useSelectedItems = (
     }
   };
 
-  // 전체선택 state
-
-  const onSelectAllClick = () => {
-    if (!all) {
-      setSelectedItems(
-        data.map(item => ({
-          cartItemId: item.id,
-          quantity: getOneItemQuantity(item.id) ?? item.quantity,
-          price: item.product.price,
-        })),
-      );
-      setAll(true);
-    } else {
-      setSelectedItems([]);
-      setAll(false);
-    }
-  };
-
   // 전체선택을 눌렀을 때 선택 해제와 전체 선택을 해주는 기능
   // 장바구니가 담긴 요소들이 모두 선택된다면 전체선택 활성화
 
@@ -100,8 +77,6 @@ const useSelectedItems = (
     selectedItems,
     onCheckboxClick,
     isSelected,
-    onSelectAllClick,
-    all,
     selectedItemQuantity,
   };
 };
